+++ tests/src/tests_ubi_write_read.c
@@ -187,7 +187,7 @@ ZTEST(ubi_write_read, one_volume_one_leb_operation_with_reboot)
 	zassert_ok(ubi_volume_create(ubi, &vol_cfg_1, &vol_id_1));
 
 	zassert_ok(ubi_device_get_info(ubi, &info_after_init));
-	zassert_equal(info_after_init.allocated_peb_count, vol_cfg_1.leb_count);
+	zassert_equal(info_after_init.reserved_peb_count, vol_cfg_1.leb_count);
 	zassert_equal(1, info_after_init.volume_count);
 
 	/* 3. Write data to LEB */
@@ -224,7 +224,7 @@ ZTEST(ubi_write_read, one_volume_one_leb_operation_with_reboot)
 
 	/* 7. Read device info */
 	zassert_ok(ubi_device_get_info(ubi, &info_after_write));
-	zassert_equal(vol_cfg_1.leb_count, info_after_write.allocated_peb_count);
+	zassert_equal(vol_cfg_1.leb_count, info_after_write.reserved_peb_count);
 	zassert_equal(1, info_after_write.volume_count);
 	zassert_equal(info_after_init.free_peb_count - 1, info_after_write.free_peb_count);
 
@@ -320,7 +320,7 @@ ZTEST(ubi_write_read, one_volume_many_leb_operations_with_reboot)
 
 	/* 6. Read device info */
 	zassert_ok(ubi_device_get_info(ubi, &info_after_write));
-	zassert_equal(vol_cfg_1.leb_count, info_after_write.allocated_peb_count);
+	zassert_equal(vol_cfg_1.leb_count, info_after_write.reserved_peb_count);
 	zassert_equal(1, info_after_write.volume_count);
 	zassert_equal(info_after_init.free_peb_count - ARRAY_SIZE(leb),
 		      info_after_write.free_peb_count);
@@ -489,7 +489,7 @@ ZTEST(ubi_write_read, many_volumes_many_leb_operations_with_reboot)
 	/* 7. Read device info */
 	zassert_ok(ubi_device_get_info(ubi, &info_after_write));
 	zassert_equal(vol_cfg_1.leb_count + vol_cfg_2.leb_count + vol_cfg_3.leb_count,
-		      info_after_write.allocated_peb_count);
+		      info_after_write.reserved_peb_count);
 	zassert_equal(3, info_after_write.volume_count);
 	zassert_equal(info_after_init.free_peb_count - vol_cfg_1.leb_count - vol_cfg_2.leb_count -
 			      vol_cfg_3.leb_count,
@@ -621,7 +621,7 @@ ZTEST(ubi_write_read, one_volume_many_lebs_io_operations_not_aligned_with_reboot
 
 	/* 7. Read device info */
 	zassert_ok(ubi_device_get_info(ubi, &info_after_write));
-	zassert_equal(vol_cfg_1.leb_count, info_after_write.allocated_peb_count);
+	zassert_equal(vol_cfg_1.leb_count, info_after_write.reserved_peb_count);
 	zassert_equal(1, info_after_write.volume_count);
 	zassert_equal(info_after_init.free_peb_count - ARRAY_SIZE(leb),
 		      info_after_write.free_peb_count);